+++ Libraries/Components/View/ViewNativeComponent.js
@@ -10,9 +10,11 @@
 
 'use strict';
 
+const AndroidConfig = require('ViewNativeComponentAndroidConfig');
 const Platform = require('Platform');
 const ReactNative = require('ReactNative');
 
+const verifyComponentAttributeEquivalence = require('verifyComponentAttributeEquivalence');
 const requireNativeComponent = require('requireNativeComponent');
 const ReactNativeViewConfigRegistry = require('ReactNativeViewConfigRegistry');
 
@@ -21,8 +23,11 @@ import type {ViewProps} from 'ViewPropTypes';
 type ViewNativeComponentType = Class<ReactNative.NativeComponent<ViewProps>>;
 
 let NativeViewComponent;
+if (Platform.OS === 'android') {
+  if (__DEV__) {
+    verifyComponentAttributeEquivalence('RCTView', AndroidConfig);
+  }
 
-if (Platform.OS === 'Android') {
   NativeViewComponent = ReactNativeViewConfigRegistry.register('RCTView', () =>
     require('ViewNativeComponentAndroidConfig'),
   );
